Relative risk regions for gameboard
Regions 3 and 5 are coveted real estate on the Reversi gameboard,
because it's harder for opponents to threaten your pieces. Especially
so for Region 5. These two regions also make it easy to take large
numbers of opponent pieces at once. So the AI in this example is
encouraged to place pieces in these regions whenever possible.

In contrast, Regions 2 and 4 are risky territory because opponents gain
access to Regions 3 and 5 after taking pieces place here. So, the AI is
discouraged from occupying these two regions.

*This information obtained from
"http://mnemstudio.org/game-reversi-example-2.htm"